+++ filen-rclone-wrapper/src/serve.rs
@@ -25,47 +25,51 @@ pub struct BasicServerDetails {
 	pub process: Child,
 }
 
+/// Starts a "webdav", "ftp", "sftp", "http" or "s3" server over Rclone.
+/// "s3" server options take the Access Key ID as `options.user` and Secret Access Key as `options.password`.
 pub async fn start_basic_server(
 	client: &Client,
 	config_dir: &Path,
 	server_type: &str,
 	options: BasicServerOptions,
 ) -> Result<BasicServerDetails> {
-	if !["webdav", "ftp", "sftp", "http"].contains(&server_type) {
+	if !["webdav", "ftp", "sftp", "http", "s3"].contains(&server_type) {
 		return Err(anyhow!("Unsupported server type: {}", server_type));
 	}
 	let url = options.url.unwrap_or("127.0.0.1:8080".to_string());
-	let auth: Option<BasicAuthentication> = if options.user.is_some() || options.password.is_some()
-	{
-		Some(BasicAuthentication {
-			user: options.user.unwrap_or("user".to_string()),
-			password: options.password.unwrap_or("password".to_string()),
-		})
-	} else {
-		None
-	};
+	let remote_str = format!("filen:{}", options.root.unwrap_or("".to_string()));
+	let mut args = vec![
+		"serve",
+		server_type,
+		&remote_str,
+		"--addr",
+		&url,
+		if options.read_only { "--read-only" } else { "" },
+	];
+	let (user, password) = (
+		options.user.clone().unwrap_or("user".to_string()),
+		options.password.clone().unwrap_or("password".to_string()),
+	);
+	let auth_key_str = format!("{},{}", user, password);
+	if server_type == "s3" {
+		args.extend(["--auth-key", &auth_key_str]);
+	} else if options.user.is_some() || options.password.is_some() {
+		args.extend(["--user", &user, "--pass", &password]);
+	}
 	let (process, _) = RcloneInstallation::initialize(client, config_dir)
 		.await?
-		.execute_in_background(&[
-			"serve",
-			server_type,
-			&format!("filen:{}", options.root.unwrap_or("".to_string())),
-			"--addr",
-			&url,
-			if auth.is_some() { "--user" } else { "" },
-			if let Some(auth) = &auth {
-				auth.user.as_str()
-			} else {
-				""
-			},
-			if auth.is_some() { "--pass" } else { "" },
-			if let Some(auth) = &auth {
-				auth.password.as_str()
-			} else {
-				""
-			},
-			if options.read_only { "--read-only" } else { "" },
-		])
+		.execute_in_background(&args)
 		.await?;
-	Ok(BasicServerDetails { url, auth, process })
+	Ok(BasicServerDetails {
+		url,
+		auth: if server_type == "s3" || (options.user.is_some() || options.password.is_some()) {
+			Some(BasicAuthentication {
+				user: user.clone(),
+				password: password.clone(),
+			})
+		} else {
+			None
+		},
+		process,
+	})
 }